Non-working broadside commit
There's a consistency issue when I generate
the memory contents randomly and then instruct
the PRUs to transfer it via the broadside and
inspect it on the other side. What's strange is
that there's a latency of 1 program execution
between the values getting reflected. See the gist
wherein the program was executed quickly within a
second, as the random seed is updated every sec, so
the mem contents match.

See: https://gist.github.com/abhishek-kakkar/a060a5d03233546cea6b

Signed-off-by: Kumar Abhishek <abhishek@theembeddedkitchen.net>